requirements.txt
einops
timm
numpy
opencv-python
nvidia-ml-py
pyyaml
Pillow
onnx
onnxruntime-gpu<1.21